rasterization optimization, rasterized polygons working
